[chore #156406804] Refactor tests and views
