+++ tests/test_Book_Model.py
@@ -5,7 +5,7 @@ from unittest import main
 import json
 
 from tests import TestBase
-from app.models.book import Book
+from app.models import Book
 
 from app.views import BOOKS
