+++ tests/test_User_Model.py
@@ -6,7 +6,7 @@ from unittest import main
 import json
 
 from tests import TestBase
-from app.models.user import User
+from app.models import User
 from app.views import USERS
